GROOVY-12355: STC: infer division and group arithmetic from the runtime's number categories
The runtime divides with floating-point math when either operand is a
Float or Double, wrapper or primitive, and with BigDecimal math for
integral, BigInteger and BigDecimal operands. The type checker only
recognised primitive float and double, so Double / Integer was inferred
as BigDecimal, and an operand typed as Number, whose runtime category
is unknown, was inferred as BigDecimal too. Statically compiled code
then cast the Double the runtime produced and failed.

Unwrap before testing the floating category, and infer BigDecimal only
when both operands belong to a category that divides that way; a
Number operand yields Number. The group operations had the same blind
spot from the other side: they took the category of whichever operand
they recognised, so Number * int was int. An operand of unknown
category now yields Number unless a floating-point partner decides the
result, as it does at runtime.

An arithmetic result that could not be typed before may now be a
Number where it was an int or BigDecimal, so code that assigned it to a
narrower type needs an explicit conversion, as the GROOVY-5539 test now
shows. StringUtil.bar with Double arguments works without change.

+++ src/test/groovy/org/codehaus/groovy/classgen/asm/sc/BugsStaticCompileTest.groovy
@@ -134,12 +134,12 @@ final class BugsStaticCompileTest extends BugsSTCTest implements StaticCompilati
         '''
     }
 
-    // GROOVY-
+    // GROOVY-5539
     @Test
     void testPowerShouldNotThrowVerifyError() {
         assertScript '''
             int squarePlusOne(int num) {
-                num ** num + 1
+                (num ** num + 1).intValue() // the sum of a power (a Number) and an int is a Number (GROOVY-12355)
             }
             assert squarePlusOne(2) == 5
         '''
